requirements.txt
flask
flask-cors
requests
face_recognition
opencv-python
numpy
Pillow
Werkzeug